Make random its own module
